requirements.txt
requests
pydantic
loguru
python-dotenv
fastapi
uvicorn


liteai
python-snippets
streamlit
mem0ai
sqlalchemy
click
pydub